+++ website/next-sitemap.config.js
@@ -1,5 +1,0 @@
-/** @type {import('next-sitemap').IConfig} */
-module.exports = {
-  siteUrl: 'https://ark-ui.com',
-  generateRobotsTxt: true,
-}