+++ website/src/components/shared/Link.tsx
@@ -1,9 +1,0 @@
-import { styled, type HTMLStyledProps } from '@/panda/jsx'
-import { link, type LinkVariantProps } from '@/panda/recipes'
-
-export type LinkProps = HTMLStyledProps<'a'> & LinkVariantProps
-
-export const Link = (props: LinkProps) => {
-  const { variant, ...rest } = props
-  return <styled.a className={link({ variant })} {...rest} />
-}