+++ packages/react/CHANGELOG.md
@@ -6,7 +6,11 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
-## [0.15.0-beta.0] - 2023-09-14
+### Changed
+
+- Add memoization to `Select` and `Combobox` item collection to improve performance.
+
+## [0.15.0] - 2023-09-14
 
 ### Added
 
@@ -224,7 +228,7 @@ description: All notable changes to this project will be documented in this file
 - Add `Toast`
 - Add `Tooltip`
 
-[unreleased]: https://github.com/chakra-ui/ark/compare/@ark-ui/react@0.15.0-beta.0...HEAD
+[unreleased]: https://github.com/chakra-ui/ark/compare/@ark-ui/react@0.15.0...HEAD
 [0.1.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.1.0
 [0.2.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.2.0
 [0.3.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.3.0
@@ -252,5 +256,5 @@ description: All notable changes to this project will be documented in this file
 [0.13.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.13.0
 [0.13.1]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.13.1
 [0.14.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.14.0
-
 [0.15.0-beta.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.15.0-beta.0
+[0.15.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.15.0